BUGFIX/MEDIUM(oio-event-agent): Prevent 0 workers in event agent
 ##### SUMMARY

Currently (SDS 5.0.0.0b1), if a worker value is set in the event agent
it will always be used. This is an issue when the server has 1 CPU
(int(1/2) = 0 workers), meaning the event agent doesn't spawn any
workers. This workaround prevents this faulty behavior until it has been
properly fixed in the code.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION